Dev UI: Change bug report to be a popover
