+++ extensions/vertx-http/deployment/src/main/java/io/quarkus/devui/deployment/ReportIssuesProcessor.java
@@ -1,0 +1,12 @@
+package io.quarkus.devui.deployment;
+
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.devui.runtime.reportissues.ReportIssuesJsonRPCService;
+import io.quarkus.devui.spi.JsonRPCProvidersBuildItem;
+
+public class ReportIssuesProcessor {
+    @BuildStep
+    JsonRPCProvidersBuildItem registerJsonRpcService() {
+        return new JsonRPCProvidersBuildItem("report-issues", ReportIssuesJsonRPCService.class);
+    }
+}
